Develop

+++ dndEncounterCalculator/dndEncounterCalculator/calculate.h
@@ -4,4 +4,4 @@
 
 using namespace std;
 
-string calculatedDifficulty(int numPlayers, int playerLevel, vector<pair<int, int>> *allMonsters);
+string calculatedDifficulty(int numPlayers, int playerLevel, vector<pair<double, int>> *allMonsters);
